requirements.txt
aiogram
python-binance